Fix typechecking for Enums.
The only problem now is that the interface is emitted with an extra
pointer, but the emitter doesn't currently have enough information to
tell the difference.

I'm thinking that it might be good to create an intermediate Go
representation that it can be compiled to before being emitted (and
possibly also a stage in the original form where all the type
information is known).